move mensa plan to a page and change plan loading
