+++ components/basic-utilities-list.vue
@@ -26,7 +26,10 @@
       </v-list-tile-content>
     </v-list-tile>
 
-    <v-list-tile @click="mensaPlanDialogOpen = true; trackMatomoEvent('Menu', 'openMensaPlanDialog')">
+    <v-list-tile 
+      to="/mensa"
+      nuxt
+      @click="trackMatomoEvent('Menu', 'openMensaPlanDialog')">
       <v-list-tile-action>
         <v-icon>fastfood</v-icon>
       </v-list-tile-action>
@@ -38,23 +41,19 @@
     </v-list-tile>
 
     <custom-timetable-dialog v-model="customTimetableDialogOpen" />
-    <mensa-plan-dialog v-model="mensaPlanDialogOpen" />
   </v-list>
 </template>
 
 <script>
 import CustomTimetableDialog from './custom-timetable-dialog.vue';
-import MensaPlanDialog from './mensa-plan-dialog.vue';
 
 export default {
   components: {
     CustomTimetableDialog,
-    MensaPlanDialog,
   },
   data() {
     return {
       customTimetableDialogOpen: false,
-      mensaPlanDialogOpen: false,
     };
   },
   methods:{
