+++ store/mensa.js
@@ -12,12 +12,21 @@ export const mutations = {
 
 export const actions = {
   async loadWeek({ state, commit }) {
+    let openDays = [];
     let weekdays = [];
     let result = [];
 
-    for(let i = 0; i<5; i++) {
-        weekdays.push(moment().startOf('isoWeek').add(i,'days'));
+    try {
+      const response = await this.$axios.get(`https://openmensa.org/api/v2/canteens/166/days`);
+      openDays = response.data;
+    } catch (error) {
+        console.error('error during Mensa API call', error.message);
     }
+
+    for(let i=0; i<3; i++) {
+        weekdays.push(moment(openDays[i].date));
+    }
+
     await Promise.all(weekdays.map(async (day) => {
         try {
             const response = await this.$axios.get(`https://openmensa.org/api/v2/canteens/166/days/${day.format('YYYY-MM-DD')}/meals`);
